refactor: move snapshot content processing to separated file
